+++ src/RobotStudio.Desktop/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("RobotStudio.Desktop.Tests")]
+[assembly: InternalsVisibleTo("RobotStudio.VisualizationDiagnostics")]